feat: split links into publication and talk pages
